♻️ [Refactoring]: deployment::hook_deploy モジュール分離
- src/component/deployment/hook_deploy.rs に impl ComponentDeployment ブロック (deploy_hook_converted / rewrite_script_paths_in_json) を抽出（pub(super)）
- src/component/deployment/hook_deploy_test.rs に hook 変換デプロイ関連テスト 14 件 + sample_claude_code_hook_json ヘルパーを移動
- 挙動不変: エラー文字列 'plugin_root is required when scripts are generated' / rewrite_script_paths_in_json の戻り値型 () を維持
- output::HookConvertOutput は deployment.rs から再エクスポートしない方針に従い pub use を DeploymentOutput のみに整理
- 既存 N=59 件 green 維持
